Updated: fetch balance from BE api
